% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{addCustomContents}
\alias{addCustomContents}
\title{Add custom contents to the tooltips}
\usage{
addCustomContents(data, callback)
}
\description{
For each row of the plot data, applies a callback function that returns
an HTML character string to be appended to the contents.
}
