% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{as_trans}
\alias{as_trans}
\title{As trans}
\usage{
as_trans(x)
}
\arguments{
\item{x}{character string, the scale name}
}
\value{
scale object
}
\description{
Gets a proper trans object from scales package. Original function 
scales::as.trans() is not working properly when scales are in Imports
}
