% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{assignLayoutNamesToPanels}
\alias{assignLayoutNamesToPanels}
\title{Assign layout names to panel numbers}
\usage{
assignLayoutNamesToPanels(gtree)
}
\arguments{
\item{gtree}{A gtree object.}
}
\value{
A character vector.
}
\description{
Returns a map of grid layout names associated with panel numbers
(e.g., on a trellis plot).
}
