% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{correctMargins}
\alias{correctMargins}
\title{Correct margin sizes}
\usage{
correctMargins(sizes, totalSize)
}
\arguments{
\item{sizes}{Numeric.}

\item{totalSize}{Numeric.}
}
\value{
Corrected `sizes`.
}
\description{
Corrects a vector of sizes for margin size.
}
