% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{filterGeoms}
\alias{filterGeoms}
\title{Filter geometries by class(es)}
\usage{
filterGeoms(geomList, classes)
}
\arguments{
\item{geomList}{A list of grobs.}

\item{classes}{A character string with geometry names.}
}
\value{
A filtered list of grobs.
}
\description{
Filter geometries by class(es)
}
