% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{freezeFactorLevels}
\alias{freezeFactorLevels}
\title{Freeze factor levels}
\usage{
freezeFactorLevels(df, characterToFactor = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{characterToFactor}{If TRUE, converts strings to factors.}
}
\value{
A data frame.
}
\description{
Reorders factor levels in a data frame, and - optionally -
converts character columns to factors.
}
