% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getAbsoluteX}
\alias{getAbsoluteX}
\title{Get absolute X position(s) of an element}
\usage{
getAbsoluteX(gtree, element, layoutName, colWidths, unit = "mm")
}
\arguments{
\item{gtree}{A gtree object.}

\item{element}{A grob.}

\item{layoutName}{Layout name (character).}

\item{colWidths}{Widths of all grid columns.}

\item{unit}{Unit (character).}
}
\description{
Get absolute X position(s) of an element
}
