% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getAbsoluteY}
\alias{getAbsoluteY}
\title{Get absolute Y position(s) of an element}
\usage{
getAbsoluteY(gtree, element, layoutName, rowHeights, unit = "mm")
}
\arguments{
\item{gtree}{A gtree object.}

\item{element}{A grob.}

\item{layoutName}{Layout name (character).}

\item{unit}{Unit (character).}

\item{colWidths}{Widths of all grid columns.}
}
\description{
Get absolute Y position(s) of an element
}
