% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getCharWidth}
\alias{getCharWidth}
\title{Get width of a proportional character}
\usage{
getCharWidth(char)
}
\arguments{
\item{char}{character string of length 1 (single character).}
}
\value{
numeric (floating-point) value.
}
\description{
Estimates how the character compares to a "standard character", like
\code{a} or \code{s}.
Returns a single floating-point value: 1 for characters of standard
width, >1 for wider characters, <1 for narrower characters.
}
