% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getDependencies}
\alias{getDependencies}
\title{load Front-End dependencies (use this function in any rendered content and
it will the files only ones)}
\usage{
getDependencies()
}
\description{
load Front-End dependencies (use this function in any rendered content and
it will the files only ones)
}
