% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getGeomCoordsForGrob}
\alias{getGeomCoordsForGrob}
\title{Get coords of grob's geometries}
\usage{
getGeomCoordsForGrob(gtree, layoutName, geomList, geom, colWidths,
  rowHeights, unit = "mm")
}
\arguments{
\item{gtree}{A gtree object.}

\item{layoutName}{Layout name (character).}

\item{geomList}{A list of grobs.}

\item{geom}{Geometry name (character) or NULL.}

\item{colWidths}{Widths of all grid columns.}

\item{rowHeights}{Heights of all grid rows.}

\item{unit}{Unit (character).}
}
\description{
Get coords of grob's geometries
}
