% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getGeomsFromGrob}
\alias{getGeomsFromGrob}
\title{Get geometries for grob}
\usage{
getGeomsFromGrob(g)
}
\arguments{
\item{g}{A \link{grob}.}
}
\value{
A list of grobs.
}
\description{
Get geometries for grob
}
