% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getGrobCol}
\alias{getGrobCol}
\title{Get grob column(s)}
\usage{
getGrobCol(gtree, layoutName)
}
\arguments{
\item{gtree}{A gtree object.}

\item{layoutName}{Layout name (a character string).}
}
\value{
An integer vector.
}
\description{
Returns grid column number(s) for a grob.
}
