% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getGrobRow}
\alias{getGrobRow}
\title{Get grob row(s)}
\usage{
getGrobRow(gtree, layoutName)
}
\arguments{
\item{gtree}{A gtree object.}

\item{layoutName}{Layout name (a character string).}
}
\value{
An integer vector.
}
\description{
Returns grid row number(s) for a grob.
}
