% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getGrobSize}
\alias{getGrobSize}
\title{Get grob size}
\usage{
getGrobSize(gtree, layoutName = NULL, unit = "mm")
}
\arguments{
\item{gtree}{A gtree object.}

\item{layoutName}{Layout name (a character string). If NULL, size
for `gtree` is calculated.}

\item{unit}{Unit name (see \link{unit}).}
}
\value{
An list of two numerics.
}
\description{
Returns grob size in specified units.
}
