% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{getLayerAesthetics}
\alias{getLayerAesthetics}
\title{Get layer aesthetics}
\usage{
getLayerAesthetics(plot)
}
\arguments{
\item{layer}{A ggplot object.}
}
\value{
A list.
}
\description{
Returns a list of aesthetics for each layer in a ggplot.
}
