% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{getLayerGeom}
\alias{getLayerGeom}
\title{Get layer geometries}
\usage{
getLayerGeom(layer)
}
\arguments{
\item{layer}{A ggplot layer object.}
}
\value{
A character vector.
}
\description{
Returns grid-compatible geometry names for a ggplot layer.
}
