% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{getNamesFromVarDict}
\alias{getNamesFromVarDict}
\title{Use columns defined in variable dictionary}
\usage{
getNamesFromVarDict(df, varDict, mapping)
}
\description{
Use columns defined in variable dictionary
}
