% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{getPlotLayerData}
\alias{getPlotLayerData}
\title{Get plot layer data}
\usage{
getPlotLayerData(plot)
}
\description{
Returns list of data elements from plot layers. If plot layer data element is 
ggplot2 waiver then plot's data element is used as default.
}
