% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getStringWidth}
\alias{getStringWidth}
\title{Get width of a proportional string}
\usage{
getStringWidth(string)
}
\arguments{
\item{string}{character string.}
}
\value{
numeric (floating-point) value.
}
\description{
Returns estimated width (in "standard characters") of a string,
as if it were rendered using a proportional font.
This function is used for the purpose of scaling text containers
when the actual width of the text is not yet known (before rendering
the text).
}
