% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{getViewportName}
\alias{getViewportName}
\title{Get viewport name}
\usage{
getViewportName(gtree, layoutName)
}
\arguments{
\item{gtree}{A gtree object.}

\item{layoutName}{Layout name (a character string).}
}
\value{
A character string.
}
\description{
Get viewport name
}
