% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{gridColWidth}
\alias{gridColWidth}
\title{Grid column width}
\usage{
gridColWidth(gtree, colNum, unit = "mm")
}
\arguments{
\item{gtree}{A gtree object.}

\item{colNum}{Column number (integer).}

\item{unit}{Unit (character).}
}
\value{
A numeric.
}
\description{
Returns width (in specified units) of a grid column.
}
