% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{gridRowHeight}
\alias{gridRowHeight}
\title{Grid row height}
\usage{
gridRowHeight(gtree, rowNum, unit = "mm")
}
\arguments{
\item{gtree}{A gtree object.}

\item{unit}{Unit (character).}

\item{colNum}{Row number (integer).}
}
\value{
A numeric.
}
\description{
Returns height (in specified units) of a grid row.
}
