% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{htmlWithGivenTooltips}
\alias{htmlWithGivenTooltips}
\title{Render plot with given tooltips data}
\usage{
htmlWithGivenTooltips(svg, data, height = NA, width = NA, point.size = 10)
}
\arguments{
\item{svg}{A SVG plot object.}

\item{data}{List with tooltip data.}

\item{height}{Plot height (in inches; optional).}

\item{width}{Plot width (in inches; optional).}

\item{point.size}{Point size for calibrating hovering accuracy (optional).}
}
\description{
Render plot with given tooltips data
}
