% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{nameToGrob}
\alias{nameToGrob}
\title{Name to grob}
\usage{
nameToGrob(gtree, layoutName)
}
\arguments{
\item{gtree}{A gtree object.}

\item{layoutName}{Layout name (as a character string).}
}
\value{
A \link{grob}.
}
\description{
Returns grob instance associated with the specified layout name.
}
