% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{orderByPanels}
\alias{orderByPanels}
\title{Order by panels}
\usage{
orderByPanels(dfList)
}
\arguments{
\item{dfList}{A list of data frames.}
}
\value{
A list of data frames.
}
\description{
Orders each data frame in a list by column \code{PANEL} if it exists.
}
