% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{parseMapping}
\alias{parseMapping}
\title{Parse plot/layer mapping}
\usage{
parseMapping(mapping)
}
\arguments{
\item{mapping}{ggplot mapping.}
}
\value{
Character.
}
\description{
Converts ggplot mapping to a character vector.
}
