% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotWithTooltips}
\alias{plotWithTooltips}
\title{Helper function that can be used to render GGPlot with ggtips in renderUI}
\usage{
plotWithTooltips(
  plot,
  varDict,
  plotScales = NULL,
  callback = NULL,
  point.size = 10,
  dpi = 72,
  width = NA,
  height = NA,
  customGrob = NULL,
  ...
)
}
\description{
Helper function that can be used to render GGPlot with ggtips in renderUI
}
