% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{readSvgAndRemoveTextLength}
\alias{readSvgAndRemoveTextLength}
\title{Auxiliary function removing `textLength` and `lengthAdjust` from SVG data}
\usage{
readSvgAndRemoveTextLength(filename)
}
\description{
Auxiliary function removing `textLength` and `lengthAdjust` from SVG data
}
