% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{removeOutOfRangeData}
\alias{removeOutOfRangeData}
\title{Remove out of range data}
\usage{
removeOutOfRangeData(data, plot, built)
}
\description{
If plot has data that was filtered when specific geom was added 
it should be filtered out of data.
}
