% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{renderWithTooltips}
\alias{renderWithTooltips}
\title{Render plot with tooltips}
\usage{
renderWithTooltips(plot, varDict, plotScales = NULL, callback = NULL,
  point.size = 10, dpi = 72, width = NA, height = NA,
  customGrob = NULL, ...)
}
\arguments{
\item{plot}{A \link{ggplot} object.}

\item{varDict}{Variable dictionary in the following format:
\code{list(<variable> = <label>, ...)},
where \code{<variable>} is a valid name of a variable mapped
with \link{aes}, and \code{<label>} is a character string.
It defines the composition of information displayed in tooltips.}

\item{plotScales}{A list with two fields: x and y. Defines axis
scales (transformations) for the purpose of displaying original
values in tooltips. If NULL (default), values are displayed "as is".}

\item{callback}{Callback function for adding custom content to the tooltips
(see the example app).}

\item{point.size}{Point size for calibrating hovering accuracy (optional).}

\item{dpi}{DPI value (optional).}

\item{width}{Plot width (in inches; optional).}

\item{height}{Plot height (in inches; optional).}

\item{customGrob}{optional grob object. It allows to pass original plot with 
grob manipulations.}

\item{...}{Additional parameters passed to \link{ggsave}.}
}
\description{
Render plot with tooltips
}
\note{
Non-numeric variables are only back-transformed to their original
values if they are passed to \link{ggplot} as factors.
}
