% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{saveAndGetTooltips}
\alias{saveAndGetTooltips}
\title{Save ggplot and get tooltips}
\usage{
saveAndGetTooltips(
  filename,
  plot = last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  ...,
  varDict,
  plotScales,
  g,
  ggPlotObj = NULL,
  callback = NULL,
  addAttributes = FALSE
)
}
\arguments{
\item{plot}{ggplot object or customGrob, see "getSvgAndTooltipdata" for
more details.}

\item{varDict}{Variable dictionary in the following format:
\code{list(<variable> = <label>, ...)},
where \code{<variable>} is a valid name of a variable mapped
with \link{aes}, and \code{<label>} is a character string.
It defines the composition of information displayed in tooltips.}

\item{plotScales}{A list with two fields: x and y. Defines axis
scales (transformations) for the purpose of displaying original
values in tooltips. If NULL (default), values are displayed "as is".}

\item{g}{A gtable object compiled from the plot (see \link{arrangeGrob}).}

\item{ggPlotObj}{optional, used if plot is a customGrob.}

\item{callback}{Callback function for adding custom content to the tooltips
(see the example app).}

\item{addAttributes}{Logical parameter determinig whether extra geom
attributes should be add to tooltip object.}
}
\value{
A list.
}
\description{
Wrapper for \link{ggsave}; after saving a plot, returns an HTML-formatted
list of tooltip data (see \link{getTooltips}).
}
