% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grobs.R
\name{setFocusTo}
\alias{setFocusTo}
\title{Set focus to viewport}
\usage{
setFocusTo(gtree, layoutName)
}
\arguments{
\item{gtree}{A gtree object.}
}
\value{
See \link{seekViewport}.
}
\description{
Sets focus to a \link{viewport} associated with a layout.
}
