% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{splitLongWords}
\alias{splitLongWords}
\title{Split long words}
\usage{
splitLongWords(
  words,
  width,
  separators = "[,;:/\\\\?\\\\!\\\\-\\\\\%]",
  NAstring = "NA"
)
}
\arguments{
\item{words}{a character vector.}

\item{width}{integer; maximal width (in characters) of a word.}

\item{separators}{character string with a regular expression.}

\item{NAstring}{character string, NA value replacement
The word is cut *after* each occurrence of the pattern.}
}
\value{
an updated character vector.
}
\description{
Takes a vector of words (character strings) and splits each word
longer than the specified length (in characters) using separators
passed as a regular expression.
}
