% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wrap}
\alias{wrap}
\title{Wrap}
\usage{
wrap(x, width, by = "\\n")
}
\arguments{
\item{x}{character, values to be wrapped}

\item{width}{numeric, max number of characters}
}
\value{
Wrapped x
}
\description{
Wrap
}
