% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wrapProportionalString}
\alias{wrapProportionalString}
\title{Wrap proportional-font string}
\usage{
wrapProportionalString(string, width, by = "\\n")
}
\arguments{
\item{string}{character string.}

\item{width}{numeric; maximum width of the wrapped text.}

\item{by}{line separator (default: newline character).}
}
\value{
wrapped character string.
}
\description{
Wraps a string, taking into account different widths of characters in
text rendered with proportional fonts ('W' is much wider than 'i').
}
